# Addresses issue #18: adding `BytesReplacer` interface to allow `BytesReplacingReader` to have a customized sizing/search strategy