… ios/readers_test.go